+++ src/librptexture/fileformat/Qoi.hpp
@@ -1,0 +1,18 @@
+/***************************************************************************
+ * ROM Properties Page shell extension. (librptexture)                     *
+ * Qoi.hpp: Quite OK Image Format image reader.                            *
+ *                                                                         *
+ * Copyright (c) 2017-2025 by David Korth.                                 *
+ * SPDX-License-Identifier: GPL-2.0-or-later                               *
+ ***************************************************************************/
+
+#pragma once
+
+#include "FileFormat.hpp"
+
+namespace LibRpTexture {
+
+FILEFORMAT_DECL_BEGIN(Qoi)
+FILEFORMAT_DECL_END()
+
+}